refactor and css

+++ public/js/chooseMode.js
@@ -2,12 +2,23 @@ document.querySelector('input').addEventListener('change', function () {
     localStorage.setItem('cardsCount', this.value);
 });
 
+const errorMessage = document.querySelector('#errorMessage');
+
 document.querySelector('#defaultModeButton').addEventListener("click", () => {
     const cardsCount = localStorage.getItem('cardsCount');
-    localStorage.clear();
-    localStorage.setItem('cardsCount', cardsCount);
-    window.location.href = '/game';
+    if (cardsCount % 2 === 0){
+        localStorage.clear();
+        localStorage.setItem('cardsCount', cardsCount);
+        window.location.href = '/game';
+    } else{
+        errorMessage.style.opacity = '1';
+    }
 });
 document.querySelector('#customModeButton').addEventListener("click", () => {
-    window.location.href = '/upload'
+    const cardsCount = localStorage.getItem('cardsCount');
+    if (cardsCount % 2 === 0){
+        window.location.href = '/upload'
+    }else{
+        errorMessage.style.opacity = '1';
+    }
 });
